management of some files added pong360 namespace
