+++ Assets/Other.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c8c4b0575cd52e34299a558967d9864c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
