+++ Assets/Scripts/Pong360/Spawner.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace pong360
+{
+    public class Spawner : MonoBehaviour
+    {
+
+        void Start()
+        {
+            print("hello");
+        }
+
+
+        void Update()
+        {
+
+        }
+    }
+
+}
