+++ Assets/Scripts/Pong360/Spawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 673de2836c8673d4c9c3735d8322611a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
